backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
supabase
google-genai
pypdf
openai
